requirements.txt
transformers>=4.35.0
diffusers>=0.21.0
torch>=2.0.0
Pillow>=10.0.0